# be/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Clinic Management API",
        "title": "Clinic Management",
        "contact": {},
        "version": "1.0"
    },
    "host": "clinicly.fly.dev",
    "basePath": "/api/v1",
    "paths": {
...
        "/staff": {
            "get": {
                "description": "Get staff",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff"
                ],
                "summary": "Get staff",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Staff name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Is descending order",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staff response",
                        "schema": {
                            "$ref": "#/definitions/controllers.StaffListResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create staff",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff"
                ],
                "summary": "Create staff",
                "parameters": [
                    {
                        "description": "Staff data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.StaffRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staff response",
                        "schema": {
                            "$ref": "#/definitions/controllers.StaffResponse"
                        }
                    }
                }
            }
        },
...
        "/staff/{id}": {
            "get": {
                "description": "Get staff by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff"
                ],
                "summary": "Get staff by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Staff id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staff response",
                        "schema": {
                            "$ref": "#/definitions/controllers.StaffResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update staff",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff"
                ],
                "summary": "Update staff",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Staff id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Staff data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateStaffRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staff response",
                        "schema": {
                            "$ref": "#/definitions/controllers.StaffResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete staff",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff"
                ],
                "summary": "Delete staff",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Staff id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staff response",
                        "schema": {
                            "$ref": "#/definitions/controllers.StaffResponse"
                        }
                    }
                }
            }
        },
        "/ticket": {
            "get": {
                "description": "Get ticket",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Get ticket",
                "parameters": [
                    {
                        "enum": [
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "Ticket status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "STT",
                        "description": "Order by",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket response",
                        "schema": {
                            "$ref": "#/definitions/controllers.TicketListResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create ticket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Create ticket",
                "parameters": [
                    {
                        "description": "Ticket data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.TicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket response",
                        "schema": {
                            "$ref": "#/definitions/controllers.TicketResponse"
                        }
                    }
                }
            }
        },
...
        "/ticket/{id}": {
            "put": {
                "description": "Update ticket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Update ticket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ticket data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket response",
                        "schema": {
                            "$ref": "#/definitions/controllers.TicketResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete ticket",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Delete ticket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket response",
                        "schema": {
                            "$ref": "#/definitions/controllers.TicketResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controllers.StaffListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Staff"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "controllers.StaffRequest": {
            "type": "object",
            "required": [
                "address",
                "birth_date",
                "email",
                "full_name",
                "gender",
                "identity_card",
                "password",
                "phone_number",
                "role"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/types.Gender"
                },
                "identity_card": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/types.Role"
                },
                "salary": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/types.StaffStatus"
                }
            }
        },
        "controllers.StaffResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Staff"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "controllers.TicketListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ticket"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "controllers.TicketRequest": {
            "type": "object",
            "required": [
                "patient_id"
            ],
            "properties": {
                "patient_id": {
                    "type": "integer"
                }
            }
        },
        "controllers.TicketResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Ticket"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "controllers.UpdateStaffRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/types.Gender"
                },
                "identity_card": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/types.Role"
                },
                "salary": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/types.StaffStatus"
                }
            }
        },
        "controllers.UpdateTicketRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "$ref": "#/definitions/types.TicketStatus"
                }
            }
        },
...
        "models.Patient": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "identity_card": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
...
        "models.Staff": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "identity_card": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "salary": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
        "models.Ticket": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "number": {
                    "type": "integer"
                },
                "patient": {
                    "$ref": "#/definitions/models.Patient"
                },
                "patient_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
...
        "types.Gender": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "Male",
                "Female",
                "Other"
            ]
        },
...
        "types.Role": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "Receptionist",
                "Doctor",
                "Pharmacist",
                "Admin"
            ]
        },
        "types.StaffStatus": {
            "type": "integer",
            "enum": [
                1,
                2
            ],
            "x-enum-varnames": [
                "Working",
                "Quit"
            ]
        },
        "types.TicketStatus": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "Waiting",
                "Processing",
                "Done"
            ]
        }
    }
}